Enable saving a blank "easy pin" for users
If a user attempted to save a blank easy pin, no action was taken to
save their pin. This contradicted the page instruction "If the pin
code is left blank, you won't need a password within your
home network.".
This change instead depends on a user checking "Enable
in-network sign in with my easy pin code" to determine whether
to save their easy pin, thereby enabling a blank pin to be saved.